chore: create useLoadPlayerBundler hook and refactor
